Fix bug where ap builds but dependent packages don't pass .cfg checks
This was because ap's .cfg file looks for AIO with a c compiler, while uses
c++ to build itself. As of recently, sconsUtils uses -std=c99 mode for gcc,
in which Posix AIO is not exposed, making the check to fail when ap is built
(silently, as sconsUtils does not check for configure failiure of the
package itself (it probably should)).

First reported by Perry Gee, on lsst-data ([LSST-data] Build problem with
datarel; Thu, 18 Jul 2013 22:37:29 -0700)